+++ test/runtime/chaos_agent.go
@@ -1,0 +1,63 @@
+// Copyright 2017-2020 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package RuntimeTest
+
+import (
+	"fmt"
+
+	. "github.com/cilium/cilium/test/ginkgo-ext"
+	"github.com/cilium/cilium/test/helpers"
+
+	. "github.com/onsi/gomega"
+)
+
+var agentChaosTests = func() {
+	var vm *helpers.SSHMeta
+
+	BeforeAll(func() {
+		vm = helpers.InitRuntimeHelper(helpers.Runtime, logger)
+	})
+
+	AfterAll(func() {
+		vm.CloseSSHClient()
+	})
+
+	It("removing leftover Cilium interfaces", func() {
+		originalLinks, err := vm.Exec("sudo ip link show | wc -l").IntOutput()
+		Expect(err).Should(BeNil())
+
+		_ = vm.Exec("sudo ip link add lxc12345 type veth peer name tmp54321")
+
+		err = vm.RestartCilium()
+		Expect(err).Should(BeNil(), "restarting Cilium failed")
+
+		status := vm.Exec("sudo ip link show lxc12345")
+		status.ExpectFail("leftover interface were not properly cleaned up")
+
+		links, err := vm.Exec("sudo ip link show | wc -l").IntOutput()
+		Expect(err).Should(BeNil(), "Cannot get link layer information")
+		Expect(links).Should(Equal(originalLinks),
+			"Some network interfaces were accidentally removed!")
+	}, 300)
+
+	It("Checking for file-descriptor leak", func() {
+		threshold := 5000
+		fds, err := vm.Exec("sudo lsof -p `pidof cilium-node-monitor` -p `pidof cilium-agent` -p `pidof cilium-docker` 2>/dev/null | wc -l").IntOutput()
+		Expect(err).Should(BeNil())
+
+		Expect(fds).To(BeNumerically("<", threshold),
+			fmt.Sprintf("%d file descriptors open from Cilium processes", fds))
+	}, 300)
+}
