+++ test/runtime/connectivity.go
@@ -1,4 +1,4 @@
-// Copyright 2019 Authors of Cilium
+// Copyright 2019-2020 Authors of Cilium
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,6 +15,10 @@
 package RuntimeTest
 
 import (
+	"context"
+	"sync"
+	"time"
+
 	. "github.com/cilium/cilium/test/ginkgo-ext"
 	"github.com/cilium/cilium/test/helpers"
 	"github.com/cilium/cilium/test/helpers/constants"
@@ -453,4 +457,74 @@ var runtimeConntrackTest = func(datapathMode string) func() {
 		})
 
 	}
+}
+
+var restartChaosTest = func() {
+	var vm *helpers.SSHMeta
+
+	BeforeAll(func() {
+		vm = helpers.InitRuntimeHelper(helpers.Runtime, logger)
+	})
+
+	AfterAll(func() {
+		vm.CloseSSHClient()
+	})
+
+	It("Checking that during restart no traffic is dropped using Egress + Ingress Traffic", func() {
+		By("Installing sample containers")
+		vm.SampleContainersActions(helpers.Create, helpers.CiliumDockerNetwork)
+		vm.PolicyDelAll().ExpectSuccess("Cannot deleted all policies")
+
+		_, err := vm.PolicyImportAndWait(vm.GetFullPath(policiesL4Json), helpers.HelperTimeout)
+		Expect(err).Should(BeNil(), "Cannot install L4 policy")
+
+		areEndpointsReady := vm.WaitEndpointsReady()
+		Expect(areEndpointsReady).Should(BeTrue(), "Endpoints are not ready after timeout")
+
+		By("Starting background connection from app2 to httpd1 container")
+		ctx, cancel := context.WithCancel(context.Background())
+		defer cancel()
+		srvIP, err := vm.ContainerInspectNet(helpers.Httpd1)
+		Expect(err).Should(BeNil(), "Cannot get httpd1 server address")
+		type BackgroundTestAsserts struct {
+			res  *helpers.CmdRes
+			time time.Time
+		}
+		backgroundChecks := []*BackgroundTestAsserts{}
+		var wg sync.WaitGroup
+		wg.Add(1)
+		go func() {
+			for {
+				select {
+				default:
+					res := vm.ContainerExec(
+						helpers.App1,
+						helpers.CurlFail("http://%s/", srvIP[helpers.IPv4]))
+					assert := &BackgroundTestAsserts{
+						res:  res,
+						time: time.Now(),
+					}
+					backgroundChecks = append(backgroundChecks, assert)
+				case <-ctx.Done():
+					wg.Done()
+					return
+				}
+			}
+		}()
+		// Sleep a bit to make sure that the goroutine starts.
+		time.Sleep(50 * time.Millisecond)
+
+		err = vm.RestartCilium()
+		Expect(err).Should(BeNil(), "restarting Cilium failed")
+
+		By("Stopping background connections")
+		cancel()
+		wg.Wait()
+
+		GinkgoPrint("Made %d connections in total", len(backgroundChecks))
+		Expect(backgroundChecks).ShouldNot(BeEmpty(), "No background connections were made")
+		for _, check := range backgroundChecks {
+			check.res.ExpectSuccess("Curl from app2 to httpd1 should work but it failed at %s", check.time)
+		}
+	})
 }
